MINIFICPP-1272 - Rebase fix
